Insert Letsencrypt::Middleware dynamically
Since all we have to do is check for force_ssl I don't see why we can't
just do that inside an initialize hook and change the way the middleware
is inserted accordingly.

Maybe I'm missing something obvious. This would reduce the amount of manual
configuration required for this gem.